move hashing method to service
